requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
langchain==0.2.12
langgraph==0.2.19
langchain-google-genai>=1.0.7
google-generativeai>=0.7.2
langsmith==0.1.95
python-dotenv==1.0.1